Tidy up types and pi impl
- Get clearer on what abort means (aborted, but not killed yet)
- Simplify errors
- Split types into specific modules
