+++ src/agent2/commands.ts
@@ -1,0 +1,7 @@
+/**
+ * A Pi RPC `prompt` command — what gets written to the Pi process stdin to
+ * start an agent run.
+ */
+export type PromptCommand = { type: "prompt"; message: string };
+export type AbortCommand = { type: "abort" };
+export type PiRpcCommand = PromptCommand | AbortCommand;
